Remove old spinner implementation
