+++ bokehjs/src/lib/models/widgets/index.ts
@@ -18,7 +18,6 @@ export {Paragraph}           from "./paragraph"
 export {PasswordInput}       from "./password_input"
 export {MultiChoice}         from "./multichoice"
 export {NumericInput}        from "./numeric_input"
-export {NumericSpinner}      from "./numeric_spinner"
 export {PreText}             from "./pretext"
 export {RadioButtonGroup}    from "./radio_button_group"
 export {RadioGroup}          from "./radio_group"
