+++ tests/integration/widgets/test_spinner.py
@@ -19,7 +19,12 @@ import pytest ; pytest
 from flaky import flaky
 
 # Bokeh imports
-from bokeh._testing.util.selenium import RECORD, ActionChains, Keys
+from bokeh._testing.util.selenium import (
+    RECORD,
+    ActionChains,
+    Keys,
+    enter_text_in_element,
+)
 from bokeh.layouts import column
 from bokeh.models import (
     Circle,
@@ -46,7 +51,7 @@ def modify_doc(doc):
 
     plot.add_glyph(source, Circle(x='x', y='y'))
     plot.add_tools(CustomAction(callback=CustomJS(args=dict(s=source), code=RECORD("data", "s.data"))))
-    spinner = Spinner(low=-1, high=10, step=0.1, value=4, css_classes=["foo"])
+    spinner = Spinner(low=-1, high=10, step=0.1, value=4, css_classes=["foo"], format="0[.]0")
 
     def cb(attr, old, new):
         source.data['val'] = [old, new]
@@ -56,70 +61,123 @@ def modify_doc(doc):
     return doc
 
 
-def enter_value_in_spinner(driver, el, value, del_prev=True):
-    if del_prev:
-        el.clear()
-    actions = ActionChains(driver)
-    actions.move_to_element(el)
-    actions.click()
-    actions.send_keys(str(value))
-    actions.send_keys(Keys.ENTER)
-    actions.perform()
-
-
 @pytest.mark.selenium
-class Test_Spinner:
-    def test_display_number_input(self, bokeh_model_page) -> None:
+class Test_Spinner(object):
+
+    def test_spinner_display(self, bokeh_model_page) -> None:
         spinner = Spinner(css_classes=["foo"])
 
         page = bokeh_model_page(spinner)
 
-        el = page.driver.find_element_by_css_selector('.foo input')
-        assert el.get_attribute('type') == "number"
+        input_el = page.driver.find_element_by_css_selector('.foo input')
+        btn_up_el = page.driver.find_element_by_css_selector('.foo .bk-spin-btn-up')
+        btn_down_el = page.driver.find_element_by_css_selector('.foo .bk-spin-btn-down')
+        assert input_el.get_attribute('type') == "text"
+        assert btn_up_el.tag_name == "button"
+        assert btn_down_el.tag_name == "button"
 
         assert page.has_no_console_errors()
 
-    def test_displays_title(self, bokeh_model_page) -> None:
+    def test_spinner_display_title(self, bokeh_model_page) -> None:
         spinner = Spinner(title="title", css_classes=["foo"])
 
         page = bokeh_model_page(spinner)
 
-        el = page.driver.find_element_by_css_selector('.foo label')
-        assert el.text == "title"
-        el = page.driver.find_element_by_css_selector('.foo input')
-        assert el.get_attribute('type') == "number"
+        label_el = page.driver.find_element_by_css_selector('.foo label')
+        assert label_el.text == "title"
+        input_el = page.driver.find_element_by_css_selector('.foo input')
+        assert input_el.get_attribute('type') == "text"
 
         assert page.has_no_console_errors()
 
-    def test_input_value_min_max_step(self, bokeh_model_page) -> None:
-        spinner = Spinner(value=1, low=0, high=10, step=1, css_classes=["foo"])
+    def test_spinner_value_format(self, bokeh_model_page) -> None:
+        spinner = Spinner(value=1, low=0, high=10, step=1, css_classes=["foo"], format="0.00")
 
         page = bokeh_model_page(spinner)
 
-        el = page.driver.find_element_by_css_selector('.foo input')
+        input_el = page.driver.find_element_by_css_selector('.foo input')
 
-        assert el.get_attribute('value') == '1'
-        assert el.get_attribute('step') == '1'
-        assert el.get_attribute('max') == '10'
-        assert el.get_attribute('min') == '0'
+        assert input_el.get_attribute('value') == '1.00'
 
         assert page.has_no_console_errors()
 
-    def test_input_smallest_step(self, bokeh_model_page) -> None:
+    def test_spinner_smallest_step(self, bokeh_model_page) -> None:
         spinner = Spinner(value=0, low=0, high=1, step=1e-16, css_classes=["foo"])
         spinner.js_on_change('value', CustomJS(code=RECORD("value", "cb_obj.value")))
 
         page = bokeh_model_page(spinner)
 
-        el = page.driver.find_element_by_css_selector('.foo input')
+        input_el = page.driver.find_element_by_css_selector('.foo input')
 
-        enter_value_in_spinner(page.driver, el, 1e-16)
+        enter_text_in_element(page.driver, input_el, "0.43654644333534")
         results = page.results
-        assert float(results['value']) == 1e-16
+        assert results['value'] == 0.43654644333534
 
-        enter_value_in_spinner(page.driver, el, 0.43654644333534)
+        enter_text_in_element(page.driver, input_el, "1e-16", click=2)
         results = page.results
-        assert float(results['value']) == 0.43654644333534
+        assert results['value'] == 1e-16
+
+        assert page.has_no_console_errors()
+
+    def test_spinner_spinning_events(self, bokeh_model_page) -> None:
+        spinner = Spinner(value=0, low=0, high=1, step=0.01, css_classes=["foo"])
+        spinner.js_on_change('value', CustomJS(code=RECORD("value", "cb_obj.value")))
+
+        page = bokeh_model_page(spinner)
+
+        input_el = page.driver.find_element_by_css_selector('.foo input')
+        btn_up_el = page.driver.find_element_by_css_selector('.foo .bk-spin-btn-up')
+        btn_down_el = page.driver.find_element_by_css_selector('.foo .bk-spin-btn-down')
+
+        enter_text_in_element(page.driver, input_el, "0.5")
+        results = page.results
+        assert results['value'] == 0.5
+
+        #click btn up
+        actions = ActionChains(page.driver)
+        actions.click(on_element=btn_up_el)
+        actions.perform()
+        results = page.results
+        assert results['value'] == 0.51
+
+        #dbl click btn down
+        actions = ActionChains(page.driver)
+        actions.double_click(on_element=btn_down_el)
+        actions.perform()
+        results = page.results
+        assert results['value'] == 0.49
+
+        #arrow up
+        actions = ActionChains(page.driver)
+        actions.click(on_element=input_el)
+        actions.send_keys(Keys.ARROW_UP)
+        actions.perform()
+        results = page.results
+        assert results['value'] == 0.50
+
+        #arrow down
+        actions = ActionChains(page.driver)
+        actions.click(on_element=input_el)
+        actions.key_down(Keys.ARROW_DOWN)
+        actions.perform()
+        results = page.results
+        assert results['value'] == 0.49
+
+        #page up
+        actions = ActionChains(page.driver)
+        actions.click(on_element=input_el)
+        actions.key_down(Keys.PAGE_UP)
+        actions.perform()
+        results = page.results
+        assert results['value'] == 0.59
+
+        #page down
+        actions = ActionChains(page.driver)
+        actions.click(on_element=input_el)
+        actions.key_down(Keys.PAGE_DOWN)
+        actions.perform()
+        results = page.results
+        assert results['value'] == 0.49
 
         assert page.has_no_console_errors()
 
@@ -127,43 +185,45 @@ class Test_Spinner:
     def test_server_on_change_round_trip(self, bokeh_server_page) -> None:
         page = bokeh_server_page(modify_doc)
 
-        el = page.driver.find_element_by_css_selector('.foo input')
+        input_el = page.driver.find_element_by_css_selector('.foo input')
 
         # same value
-        enter_value_in_spinner(page.driver, el, 4)
+        enter_text_in_element(page.driver, input_el, "4", click=2)
         page.click_custom_action()
         results = page.results
         assert results['data']['val'] == ["a", "b"]
 
         # new valid value
-        enter_value_in_spinner(page.driver, el, 5)
+        enter_text_in_element(page.driver, input_el, "5", click=2)
         page.click_custom_action()
         results = page.results
         assert results['data']['val'] == [4, 5]
 
         # new overflow value
-        enter_value_in_spinner(page.driver, el, 11)
+        enter_text_in_element(page.driver, input_el, "11", click=2)
         page.click_custom_action()
         results = page.results
         assert results['data']['val'] == [5, 10]
 
         # new underflow value
-        enter_value_in_spinner(page.driver, el, -2)
+        enter_text_in_element(page.driver, input_el, "-2", click=2)
         page.click_custom_action()
         results = page.results
         assert results['data']['val'] == [10, -1]
 
         # new decimal value
-        enter_value_in_spinner(page.driver, el, 5.1)
+        input_el.clear() #negative previous values needs a triple click to be selected
+        enter_text_in_element(page.driver, input_el, "5.1")
         page.click_custom_action()
         results = page.results
-        assert results['data']['val'] == [-1, 5.1]
+        assert results['data']['val'] == [None, 5.1]
 
         # new decimal value test rounding
-        enter_value_in_spinner(page.driver, el, 5.19)
+        enter_text_in_element(page.driver, input_el, "5.19", click=2)
         page.click_custom_action()
         results = page.results
-        assert results['data']['val'] == [5.1, 5.2]
+        assert results['data']['val'] == [5.1, 5.19]
+        assert input_el.get_attribute('value') == '5.2'
 
         # XXX (bev) disabled until https://github.com/bokeh/bokeh/issues/7970 is resolved
         # assert page.has_no_console_errors()
